python-ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
httpx==0.26.0
python-multipart==0.0.6

# LLM Providers (install the one you need)
google-generativeai==0.3.2
# anthropic==0.18.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
